src/Producer_api/tests/requirements.txt
confluent-kafka[avro,schema-registry,protobuf]==2.11.0
protobuf==6.31.1
fastavro<2
requests==2.32.4
avro>=1.11.1,<2
googleapis-common-protos==1.70.0
httpx>=0.26
authlib>=1.0.0
